Added two notebooks based on yours _FL
I've added a copy of two of your notebooks in the notebook folder (_FL.ipynb).

I've added a Data folder and put one file with real data there. This is loaded in both notebooks.
I'm changing the directory in the notebook. Should work for you too and is easier for me.

Changed pool to pool_bls as keyword in test_injecting_transits_FL.ipynb.
Added pointsize to definition of function in recovery/plot.py.

(May want to change where injected_planets.csv is saved.)

Everything working so far.